+++ backend/src/test/java/com/woowacourse/momo/member/service/MemberServiceTest.java
@@ -22,6 +22,7 @@ import com.woowacourse.momo.global.exception.exception.MomoException;
 import com.woowacourse.momo.category.domain.Category;
 import com.woowacourse.momo.group.domain.group.Group;
 import com.woowacourse.momo.group.domain.group.GroupRepository;
+import com.woowacourse.momo.group.service.GroupFindService;
 import com.woowacourse.momo.member.domain.Member;
 import com.woowacourse.momo.member.domain.MemberRepository;
 import com.woowacourse.momo.member.service.dto.request.ChangeNameRequest;
@@ -46,6 +47,9 @@ class MemberServiceTest {
 
     @Autowired
     private GroupRepository groupRepository;
+
+    @Autowired
+    private GroupFindService groupFindService;
 
     private Member savedHost;
 
@@ -118,7 +122,7 @@ class MemberServiceTest {
 
         memberService.deleteById(memberId);
 
-        List<Group> groups = groupRepository.findParticipatedGroups(memberId);
+        List<Group> groups = groupFindService.findAllThatParticipated(memberId);
         assertThat(groups).isEmpty();
     }
